feature

+++ models/locationModel.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Double, DateTime, ARRAY
+from sqlalchemy import Column, Integer, String, ForeignKey, Double, DateTime, ARRAY, Float
 from database import Base
 
 
@@ -10,5 +10,5 @@ class Location(Base):
     creation_date  = Column(DateTime(timezone=True))
     country_id = Column(Integer, ForeignKey("Country.id"))
     image_id = Column(Integer, ForeignKey("Image.id"))
-    longitude = Column(Integer)
-    latitude = Column(Integer)
+    longitude = Column(Float)
+    latitude = Column(Float)
